requirements.txt
boto3==1.26.37
botocore==1.29.37
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==4.4.2
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
httplib2==0.21.0
idna==3.4
imageio==2.23.0
imageio-ffmpeg==0.4.7
jmespath==1.0.1
moviepy==1.0.3
mutagen==1.46.0
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
proglog==0.1.10
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.16.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
websockets==10.4
yt-dlp==2022.11.11
